+++ presto-main/src/test/java/com/facebook/presto/operator/aggregation/TestDoubleAverageAggregation.java
@@ -7,7 +7,7 @@ public class TestDoubleAverageAggregation
     extends AbstractTestAggregationFunction
 {
     @Override
-    public BlockCursor getSequenceCursor(long max)
+    public BlockCursor getSequenceCursor(int max)
     {
         return new DoubleSequenceCursor(max);
     }
