+++ presto-main/src/test/java/com/facebook/presto/operator/aggregation/TestLongAverageAggregation.java
@@ -7,7 +7,7 @@ public class TestLongAverageAggregation
     extends AbstractTestAggregationFunction
 {
     @Override
-    public BlockCursor getSequenceCursor(long max)
+    public BlockCursor getSequenceCursor(int max)
     {
         return new LongSequenceCursor(max);
     }
